# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for Star Mall service.",
        "title": "Star Mall API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9090",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/merchants/delete": {
            "delete": {
                "description": "删除商家",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "merchant"
                ],
                "summary": "删除商家",
                "responses": {
                    "200": {
                        "description": "merchant deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/merchants/getMerchant/{merchantID}": {
            "get": {
                "description": "根据商户 ID 获取商户信息, 返回商户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "merchant"
                ],
                "summary": "根据商户 ID 获取商户信息",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Merchant ID",
                        "name": "merchantID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Merchant details retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/domain.Merchant"
                        }
                    },
                    "400": {
                        "description": "Invalid merchant ID",
                        "schema": {
                            "$ref": "#/definitions/handler.MerchantRegisterResponse"
                        }
                    },
                    "404": {
                        "description": "Merchant not found",
                        "schema": {
                            "$ref": "#/definitions/handler.MerchantRegisterResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.MerchantRegisterResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/merchants/register": {
            "post": {
                "description": "注册商家",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "merchant"
                ],
                "summary": "注册商家",
                "parameters": [
                    {
                        "description": "Merchant registration information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.MerchantRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully registered merchant",
                        "schema": {
                            "$ref": "#/definitions/handler.MerchantRegisterResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/handler.MerchantRegisterResponse"
                        }
                    },
                    "409": {
                        "description": "Email or phone already registered",
                        "schema": {
                            "$ref": "#/definitions/handler.MerchantRegisterResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.MerchantRegisterResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/merchants/search/{merchant_name}": {
            "get": {
                "description": "根据商家名称搜索商家, 返回匹配的商家列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "merchant"
                ],
                "summary": "根据商家名称搜索商家",
                "parameters": [
                    {
                        "minLength": 2,
                        "type": "string",
                        "description": "Merchant name to search for (minimum 2 characters)",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of merchants matching the search criteria",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Merchant"
                            }
                        }
                    },
                    "204": {
                        "description": "No merchants found",
                        "schema": {
                            "$ref": "#/definitions/handler.MerchantRegisterResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid merchant name",
                        "schema": {
                            "$ref": "#/definitions/handler.MerchantRegisterResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.MerchantRegisterResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/merchants/update": {
            "patch": {
                "description": "修改商家信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "merchant"
                ],
                "summary": "修改商家信息",
                "parameters": [
                    {
                        "description": "Merchant update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.MerchantUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "merchant updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Merchant": {
            "type": "object",
            "properties": {
                "businessTypes": {
                    "description": "店铺经营类型",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "categoryId": {
                    "description": "店铺分类ID",
                    "type": "integer",
                    "example": 100
                },
                "createAt": {
                    "type": "integer",
                    "example": 1609459200
                },
                "deleteAt": {
                    "type": "integer",
                    "example": 0
                },
                "email": {
                    "type": "string",
                    "example": "contact@starelectronics.com"
                },
                "id": {
                    "type": "integer",
                    "example": 12345
                },
                "imageId": {
                    "description": "店铺图片ID",
                    "type": "integer",
                    "example": 2001
                },
                "licenseImageId": {
                    "description": "营业执照图片ID",
                    "type": "integer",
                    "example": 1001
                },
                "name": {
                    "type": "string",
                    "example": "Star Electronics"
                },
                "oldName": {
                    "description": "曾用名",
                    "type": "string",
                    "example": "Star Tech"
                },
                "phone": {
                    "type": "string",
                    "example": "+8613800138000"
                },
                "realId": {
                    "description": "身份证号",
                    "type": "string",
                    "example": "310000199001011234"
                },
                "realName": {
                    "description": "法人真实姓名",
                    "type": "string",
                    "example": "张三"
                },
                "score": {
                    "description": "商户评分",
                    "type": "number",
                    "example": 4.8
                },
                "status": {
                    "type": "integer",
                    "example": 1
                },
                "tag": {
                    "description": "商户标签",
                    "type": "integer",
                    "example": 1
                },
                "updateAt": {
                    "type": "integer",
                    "example": 1609459200
                },
                "userId": {
                    "type": "integer",
                    "example": 67890
                }
            }
        },
...
        "handler.MerchantRegisterRequest": {
            "type": "object",
            "required": [
                "businessType",
                "email",
                "name",
                "password",
                "phone",
                "realID",
                "realName",
                "verificationCode"
            ],
            "properties": {
                "businessType": {
                    "description": "BusinessType represents the categories of business the merchant operates in\nRequired: true\nExample: [1, 3, 5]",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "email": {
                    "description": "Email is the merchant's contact email address\nRequired: true\nExample: merchant@example.com",
                    "type": "string"
                },
                "name": {
                    "description": "Name is the merchant's business name\nRequired: true\nMin length: 2\nMax length: 32",
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 2
                },
                "password": {
                    "description": "Password for merchant account\nRequired: true\nMin length: 8\nFormat: password",
                    "type": "string",
                    "minLength": 8
                },
                "phone": {
                    "description": "Phone is the merchant's contact phone number in E.164 format\nRequired: true\nExample: +8613800138000",
                    "type": "string"
                },
                "realID": {
                    "description": "RealID is the ID card number of the merchant's representative\nRequired: true\nLength: 18",
                    "type": "string",
                    "maxLength": 18,
                    "minLength": 18
                },
                "realName": {
                    "description": "RealName is the legal name of the merchant's representative\nRequired: true\nMin length: 2\nMax length: 32",
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 2
                },
                "verificationCode": {
                    "description": "用户收到的邮箱验证码, 用于验证用户身份",
                    "type": "string"
                }
            }
        },
        "handler.MerchantRegisterResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "HTTP status code",
                    "type": "integer",
                    "example": 201
                },
                "email": {
                    "description": "Email 邮箱地址",
                    "type": "string",
                    "example": "merchant@example.com"
                },
                "merchantId": {
                    "description": "商家ID",
                    "type": "integer",
                    "example": 12345
                },
                "phone": {
                    "description": "Phone 商家店铺电话",
                    "type": "string",
                    "example": "+8613800138000"
                },
                "uploadToken": {
                    "description": "UploadToken 后续传输图片所需的 临时用户凭证",
                    "type": "string",
                    "example": "\u003cKEY\u003e"
                },
                "userId": {
                    "description": "商家对应的 用户ID",
                    "type": "integer",
                    "example": 67890
                }
            }
        },
        "handler.MerchantUpdateRequest": {
            "description": "商家更新请求",
            "type": "object",
            "required": [
                "businessType",
                "cateId",
                "email",
                "name",
                "phone"
            ],
            "properties": {
                "businessType": {
                    "description": "BusinessType 营业类型",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "cateId": {
                    "description": "CateID 商家店铺分类",
                    "type": "integer"
                },
                "email": {
                    "description": "Email 商家店铺邮箱",
                    "type": "string"
                },
                "name": {
                    "description": "Name 商家店铺名称\nRequired: true\nMin length: 2\nMax length: 32",
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 2
                },
                "phone": {
                    "description": "Phone 商家店铺电话",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "输入 JWT token，格式为：Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}